+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Datetimes are now offset-aware.
 
+### Fixed
+- Names of the exa tools.
+
 ## [v0.7.1] - 21.07.2025
 
 ### Added